# Week7/python-server/openapi.yaml
openapi: 3.0.0
info:
  description: A simple API for managing products and stores with basic CRUD operations.
  title: Product and Store Management API
  version: 1.0.0
servers:
- description: Local development server
  url: http://localhost:8000
paths:
  /products:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Product"
                type: array
          description: A list of products
      summary: List all products
      tags:
      - Products
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product created
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found
      summary: Delete a product
      tags:
      - Products
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product details
        "404":
          description: Product not found
      summary: Get product details
      tags:
      - Products
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product updated
        "404":
          description: Product not found
      summary: Update a product
      tags:
      - Products
  /stores:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Store"
                type: array
          description: A list of stores
      summary: List all stores
      tags:
      - Stores
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
          description: Store created
      summary: Create a new store
      tags:
      - Stores
  /stores/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Store deleted
        "404":
          description: Store not found
      summary: Delete a store
      tags:
      - Stores
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
          description: Store details
        "404":
          description: Store not found
      summary: Get store details
      tags:
      - Stores
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
          description: Store updated
        "404":
          description: Store not found
      summary: Update a store
      tags:
      - Stores
components:
  schemas:
    ProductInput:
      example:
        quantity: 50
        price: 999.99
        name: Apple iPhone 15
        description: Latest model with A16 Bionic chip.
        category: Electronics
      properties:
        name:
          example: Apple iPhone 15
          title: name
          type: string
        description:
          example: Latest model with A16 Bionic chip.
          title: description
          type: string
        price:
          example: 999.99
          format: float
          title: price
          type: number
        quantity:
          example: 50
          title: quantity
          type: integer
        category:
          example: Electronics
          title: category
          type: string
      required:
      - name
      - price
      title: ProductInput
      type: object
    Product:
      allOf:
      - $ref: "#/components/schemas/ProductInput"
      - properties:
          id:
            example: 123e4567-e89b-12d3-a456-426614174000
            format: uuid
            type: string
        type: object
      example:
        quantity: 50
        price: 999.99
        name: Apple iPhone 15
        description: Latest model with A16 Bionic chip.
        id: 123e4567-e89b-12d3-a456-426614174000
        category: Electronics
      title: Product
    StoreInput:
      example:
        phone: +1-202-555-0173
        name: Downtown Electronics
        location: "123 Main St, New York, NY"
      properties:
        name:
          example: Downtown Electronics
          title: name
          type: string
        location:
          example: "123 Main St, New York, NY"
          title: location
          type: string
        phone:
          example: +1-202-555-0173
          title: phone
          type: string
      required:
      - location
      - name
      title: StoreInput
      type: object
    Store:
      allOf:
      - $ref: "#/components/schemas/StoreInput"
      - properties:
          id:
            example: 987f6543-a21b-34c5-d678-098765432100
            format: uuid
            type: string
        type: object
      example:
        phone: +1-202-555-0173
        name: Downtown Electronics
        location: "123 Main St, New York, NY"
        id: 987f6543-a21b-34c5-d678-098765432100
      title: Store